+++ packages/@runlightyear/github/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/github
 
+## 0.13.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.21.0
+
 ## 0.13.0
 
 ### Minor Changes