+++ packages/@runlightyear/gmail/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/gmail
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.21.0
+
 ## 0.1.2
 
 ### Patch Changes